+++ src/main/java/com/shepard1992/gmail/twoya_formochka/repository/api/OrderRepository.java
@@ -1,12 +1,8 @@
 package com.shepard1992.gmail.twoya_formochka.repository.api;
 
 import com.shepard1992.gmail.twoya_formochka.repository.entity.Order;
-import com.shepard1992.gmail.twoya_formochka.repository.specification.OrderSpecification;
-import com.shepard1992.gmail.twoya_formochka.view.model.GetOrderPl;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import java.util.List;
-
-public interface OrderRepository extends JpaRepository<Order, Long>, JpaSpecificationExecutor<Order> {
+public interface OrderRepository extends JpaRepository<Order, Integer>, JpaSpecificationExecutor<Order> {
 }
