+++ src/main/java/com/shepard1992/gmail/twoya_formochka/view/model/FilterOrderPl.java
@@ -3,13 +3,11 @@ package com.shepard1992.gmail.twoya_formochka.view.model;
 import lombok.Builder;
 import lombok.Data;
 
-import java.time.ZonedDateTime;
-
 @Data
 @Builder
 public class FilterOrderPl {
 
-    private Long id;
+    private Integer id;
 
     private String firstName;
